+++ CHANGELOG.md
@@ -11,6 +11,12 @@ The format is based on [Keep a Changelog], and this project adheres to
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
+## [Unreleased]
+
+### Added
+
+- Added `stubs.UUIDSequence.Count()` method.
+
 ## [0.19.19] - 2025-12-18
 
 ### Added